refactor: existWinnerをリフトアップ
