+++ src/App.tsx
@@ -1,8 +1,9 @@
-import { useState } from "react";
+import { useState, useMemo } from "react";
 import { Board } from "./components/Board";
 import { GameStatus } from "./components/GameStatus";
 import { PlayHistory } from "./components/PlayHistory";
 import type { History, BoardType } from "./type";
+import { calculateWinner } from "./components/model";
 
 export default function Game(): JSX.Element {
   const [playHistory, setPlayHistory] = useState<History>([
@@ -11,8 +12,20 @@ export default function Game(): JSX.Element {
   const [currentPlayHistoryIndex, setCurrentPlayHistoryIndex] =
     useState<number>(0);
 
-  const xIsNext: boolean = currentPlayHistoryIndex % 2 === 0;
-  const currentBoard: BoardType = playHistory[currentPlayHistoryIndex];
+  const xIsNext: boolean = useMemo(
+    () => currentPlayHistoryIndex % 2 === 0,
+    [currentPlayHistoryIndex]
+  );
+
+  const currentBoard: BoardType = useMemo(
+    () => playHistory[currentPlayHistoryIndex],
+    [playHistory, currentPlayHistoryIndex]
+  );
+
+  const existWinner: number[][] | null = useMemo(
+    () => calculateWinner(currentBoard),
+    [currentBoard]
+  );
 
   function handlePlayHistory(latestBoard: BoardType): void {
     setPlayHistory((prevPlayHistory) => {
@@ -28,11 +41,16 @@ export default function Game(): JSX.Element {
   return (
     <div className="game">
       <div className="game-board">
-        <GameStatus board={currentBoard} xIsNext={xIsNext} />
+        <GameStatus
+          board={currentBoard}
+          xIsNext={xIsNext}
+          existWinner={existWinner}
+        />
         <Board
           xIsNext={xIsNext}
           board={currentBoard}
           onPlay={handlePlayHistory}
+          existWinner={existWinner}
         />
       </div>
       <div className="game-info">
